refactored quick_cert and validate out
