+++ crate/cli/src/actions/symmetric/keys/mod.rs
@@ -5,9 +5,7 @@ use self::{
     create_key::CreateKeyAction, destroy_key::DestroyKeyAction, revoke_key::RevokeKeyAction,
 };
 use crate::{
-    actions::shared::{
-        ExportKeyAction, GetAttributesAction, ImportKeyAction, UnwrapKeyAction, WrapKeyAction,
-    },
+    actions::shared::{ExportKeyAction, ImportKeyAction, UnwrapKeyAction, WrapKeyAction},
     error::CliError,
 };
 
@@ -25,7 +23,6 @@ pub enum KeysCommands {
     Unwrap(UnwrapKeyAction),
     Revoke(RevokeKeyAction),
     Destroy(DestroyKeyAction),
-    GetAttributes(GetAttributesAction),
 }
 
 impl KeysCommands {
@@ -38,7 +35,6 @@ impl KeysCommands {
             Self::Unwrap(action) => action.run(kms_rest_client).await?,
             Self::Revoke(action) => action.run(kms_rest_client).await?,
             Self::Destroy(action) => action.run(kms_rest_client).await?,
-            Self::GetAttributes(action) => action.run(kms_rest_client).await?,
         };
 
         Ok(())
