+++ crate/cli/src/tests/certificates/mod.rs
@@ -1,10 +1,10 @@
-pub mod certify;
+mod certify;
 pub mod encrypt;
 mod export;
+mod get_attributes;
 pub mod import;
 pub mod openssl;
-
-mod csr;
-mod get_attributes;
+pub mod quick_cert;
+mod validate;
 
 pub(crate) const SUB_COMMAND: &str = "certificates";
